# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Node_JS_Training -> User module",
        "version": "1.0.0",
        "description": "API documentation for user module"
    },
    "servers": [
        {
            "url": "http://localhost:3000/user",
            "description": "Local server"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/update-profile": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user profile",
                "description": "Updates the profile information of an authenticated user.",
                "operationId": "updateUserProfile",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "JohnDoe",
                                        "description": "The username of the user"
                                    },
                                    "useremail": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com",
                                        "description": "The email address of the user"
                                    }
                                },
                                "additionalProperties": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "user": {
                                            "type": "object",
                                            "properties": {
                                                "userId": {
                                                    "type": "string",
                                                    "example": "12345"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "JohnDoe"
                                                },
                                                "useremail": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                }
                                            },
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized access"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}